move courses API to file
