+++ src/containers/userMyPage.js
@@ -7,7 +7,7 @@ import { BeatLoader } from "react-spinners";
 import setPageMeta from "utils/seo";
 import CourseInfoRow from "components/CourseInfoRow";
 import { useUserData } from "components/Providers/UserProvider";
-import { useCourseSearchingContext } from "components/Providers/CourseSearchingProvider";
+import { fetchFavoriteCourses } from "queries/course";
 import { useDisplayTags } from "components/Providers/DisplayTagsProvider";
 import { useCourseTable } from "components/Providers/CourseTableProvider";
 import { fetchUserById } from "queries/user";
@@ -17,7 +17,6 @@ function UserMyPage() {
   const { setUser, user: userInfo } = useUserData();
   const toast = useToast();
   const { user, isLoading, getAccessTokenSilently } = useAuth0();
-  const { searchError, fetchFavoriteCourses } = useCourseSearchingContext();
   const { courseTable, setCourseTable } = useCourseTable();
   const { displayTags } = useDisplayTags();
   const [favorite_list, setFavorite_list] = useState([]);
